moving ui related utilities to utils/ui file
